move filterChain and ResourceLocation
