+++ avaje-jex/src/main/java/io/avaje/jex/HttpFilter.java
@@ -2,14 +2,16 @@ package io.avaje.jex;
 
 import java.io.IOException;
 
+import com.sun.net.httpserver.HttpExchange;
+
 /**
  * A filter used to pre- and post-process incoming requests. Pre-processing occurs before the
  * application's exchange handler is invoked, and post-processing occurs after the exchange handler
  * returns. Filters are organized in chains, and are associated with {@link Context} instances.
  *
  * <p>Each {@code HttpFilter} in the chain, invokes the next filter within its own {@link
- * #filter(Context, FilterChain)} implementation. The final {@code HttpFilter} in the chain invokes the
- * applications exchange handler.
+ * #filter(Context, FilterChain)} implementation. The final {@code HttpFilter} in the chain invokes
+ * the applications exchange handler.
  */
 @FunctionalInterface
 public interface HttpFilter {
@@ -36,4 +38,22 @@ public interface HttpFilter {
    * @param chain the {@code FilterChain} which allows the next filter to be invoked
    */
   void filter(Context ctx, FilterChain chain) throws IOException;
+
+  /**
+   * Filter chain that contains all subsequent filters that are configured, as well as the final
+   * route.
+   */
+  @FunctionalInterface
+  public interface FilterChain {
+
+    /**
+     * Calls the next filter in the chain, or else the user's exchange handler, if this is the final
+     * filter in the chain. The {@link HttpFilter} may decide to terminate the chain, by not calling
+     * this method. In this case, the filter <b>must</b> send the response to the request, because
+     * the application's {@linkplain HttpExchange exchange} handler will not be invoked.
+     *
+     * @throws IOException if an I/O error occurs
+     */
+    void proceed() throws IOException;
+  }
 }
