--- avaje-jex/src/main/java/io/avaje/jex/ResourceLocation.java
@@ -1,6 +1,0 @@
-package io.avaje.jex;
-
-public enum ResourceLocation {
-  CLASS_PATH,
-  FILE
-}
