+++ avaje-jex/src/test/java/io/avaje/jex/compression/CompressionTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Test;
 
 import io.avaje.jex.Jex;
-import io.avaje.jex.ResourceLocation;
+import io.avaje.jex.StaticContentConfig.ResourceLocation;
 import io.avaje.jex.core.Constants;
 import io.avaje.jex.jdk.TestPair;
